ggstats: Extension to 'ggplot2' for Plotting Stats

Provides new statistics, new geometries and new positions for 'ggplot2' and a suite of functions to facilitate the creation of statistical plots.

See also

Author

Maintainer: Joseph Larmarange joseph@larmarange.net (ORCID)